[ClockPicker] Move clock dimensions to the module where they're used
